+++ Changelog.md
@@ -5,12 +5,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+
+
+## [Unreleased]
+
+## [1.0.1] - 2020-11-03
+
+### Added
+- reactotron.
+- core navigation app.
+- theme.
+- i18n.
+- blank screens.
+- basic component.
+
+
 ## [Unreleased]
 
 ## [1.0.0] - 2020-11-03
 
 ### Added
-
 - Expo SDK 39.
 - Version navigation.
 - initial step for Storybook.